added if to gh-pages

+++ scripts/remote.deploy.sh
@@ -1,14 +1,20 @@
 #!/bin/bash
-set -o errexit
+BRANCH="project"
 
-git config --global user.email "travis@travis-ci.org"
-git config --global user.name "Travis CI"
+if [ "$TRAVIS_BRANCH" == "$BRANCH" ]; then
+	set -o errexit
 
-cp README.md build/_site/web
-cp LICENSE build/_site/web/license.txt
+	git config --global user.email "travis@travis-ci.org"
+	git config --global user.name "Travis CI"
 
-cd build/_site/web
-git init
-git add --a
-git commit -m "Deploy to Github Pages"
-git push --force --quiet "https://${GITHUB_TOKEN}@github.com/${GITHUB_REPO}.git" master:gh-pages
+	cp README.md build/_site/web
+	cp LICENSE build/_site/web/license.txt
+
+	cd build/_site/web
+	git init
+	git add --a
+	git commit -m "Deploy to Github Pages"
+	git push --force --quiet "https://${GITHUB_TOKEN}@github.com/${GITHUB_REPO}.git" master:gh-pages
+else
+	echo "Branch is not $BRANCH. Skipping build"
+fi
